requirements.txt
streamlit
pandas
plotly
gspread
Pillow